+++ src/cg/lex.cpp
@@ -786,7 +786,7 @@ namespace viua {
                                                 "0.0");
                         }
                         continue;
-                    } else if (token == "strstore") {
+                    } else if (token == "string") {
                         tokens.push_back(token);                 // mnemonic
                         tokens.push_back(input_tokens.at(++i));  // target register
                         if (not is_register_set_name(input_tokens.at(i + 1))) {
@@ -1457,7 +1457,7 @@ namespace viua {
                                                 "0.0");
                         }
                         continue;
-                    } else if (token == "strstore") {
+                    } else if (token == "string") {
                         tokens.push_back(input_tokens.at(++i));  // target register
                         if (not is_register_set_name(input_tokens.at(i + 1))) {
                             tokens.emplace_back(input_tokens.at(i).line(), input_tokens.at(i).character(),
@@ -2388,7 +2388,7 @@ namespace viua {
                         tokens.back().original("default");
                         continue;
                     }
-                    if (match(input_tokens, i, {"strstore", "", "default"})) {
+                    if (match(input_tokens, i, {"string", "", "default"})) {
                         tokens.push_back(input_tokens.at(++i));  // push target register token
                         ++i;
                         tokens.emplace_back(input_tokens.at(i).line(), input_tokens.at(i).character(),